backend/requirements.txt
fastapi
uvicorn[standard]
torch
transformers
pillow
python-multipart

